+++ CHANGELOG.md
@@ -1,12 +1,27 @@
+## 2022-07-08
+
+### Added
+
+- Added `RemoteSignal:wait`.
+
+### Changed
+
+- Removed `RemoteSignal:disonnectAll` and `ClientRemoteSignal:disonnectAll`.
+- Removed `RemoteSignal:connectOnce` and `ClientRemoteSignal:connectOnce`.
+- Fix bug with `ClientRemoteSignal:connect` and `ClientRemoteSignal:wait` not receiving the dispatched events properly in some edge cases.
+- `RemoteSignal:connect` and `ClientRemoteSignal:connect` now return an `RBXScriptConnectionObject`. 
+- Internal code refactor and improvements.
+- Documentation improvements.
+
 ## 2022-07-07
 
 ### Added
 
-- Added `ReemoteSignal.fireAllClientsExcept`.
+- Added `RemoteSignal:fireAllClientsExcept`.
 
 ### Changed
 
-- Change `RemoteSignal.fireForClients` to `RemoteSignal.fireClients`.
+- Change `RemoteSignal:fireForClients` to `RemoteSignal:fireClients`.
 - Internal code refactor for `network`.
 - Documentation improvements for `network`.